+++ ControleDespesas_ApiNetCore21_Dapper/ControleDespesas.Domain/Usuarios/Interfaces/Handlers/IUsuarioHandler.cs
@@ -7,8 +7,8 @@ namespace ControleDespesas.Domain.Usuarios.Interfaces.Handlers
     public interface IUsuarioHandler
     {
         ICommandResult<Notificacao> Handler(AdicionarUsuarioCommand command);
-        ICommandResult<Notificacao> Handler(int id, AtualizarUsuarioCommand command);
-        ICommandResult<Notificacao> Handler(int id);
+        ICommandResult<Notificacao> Handler(long id, AtualizarUsuarioCommand command);
+        ICommandResult<Notificacao> Handler(long id);
         ICommandResult<Notificacao> Handler(LoginUsuarioCommand command);
     }
 }
