+++ ControleDespesas_ApiNetCore21_Dapper/ControleDespesas.Test/AppConfigurations/QueriesSQL/CREATE.sql
@@ -47,7 +47,7 @@ DROP PROCEDURE [dbo].[ELMAH_LogError]
 IF NOT EXISTS (SELECT 1 FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_TYPE='BASE TABLE' AND TABLE_NAME='Usuario') 
 BEGIN
 CREATE TABLE [dbo].[Usuario](
-	[Id] [int] IDENTITY(1,1) NOT NULL,
+	[Id] [bigint] IDENTITY(1,1) NOT NULL,
 	[Login] [nvarchar](50) NOT NULL,
 	[Senha] [nvarchar](15) NOT NULL,
 	[Privilegio] [int] NOT NULL,
@@ -96,7 +96,7 @@ IF NOT EXISTS (SELECT 1 FROM INFORMATION_SCHEMA.TABLES WHERE TABLE_TYPE='BASE TA
 BEGIN
 	CREATE TABLE [dbo].[Pessoa](
 		[Id] [int] IDENTITY(1,1) NOT NULL,
-		[IdUsuario] [int] NOT NULL,
+		[IdUsuario] [bigint] NOT NULL,
 		[Nome] [nvarchar](100) NOT NULL,
 		[ImagemPerfil] [text] NOT NULL,
 		CONSTRAINT [PK_Pessoa] PRIMARY KEY CLUSTERED
